+++ source/utility/BackGround.cpp
@@ -123,7 +123,7 @@ int BackGround::setBmpXY( int arg_bmpNo , float arg_x , float arg_y )
 /*/
 /*	切り取り位置のセット
 /*/
-int BackGround::setBmpUV( int arg_bmpNo , int arg_u , int arg_v )
+int BackGround::setBmpUV( int arg_bmpNo , float arg_u , float arg_v )
 {
 	bmpBGTable_[ arg_bmpNo ]._u = arg_u ;
 	bmpBGTable_[ arg_bmpNo ]._v = arg_v ;
@@ -134,7 +134,7 @@ int BackGround::setBmpUV( int arg_bmpNo , int arg_u , int arg_v )
 /*/
 /*	幅、高さのセット
 /*/
-int BackGround::setBmpWH( int arg_bmpNo , int arg_w , int arg_h )
+int BackGround::setBmpWH( int arg_bmpNo , float arg_w , float arg_h )
 {
 	bmpBGTable_[ arg_bmpNo ]._w = arg_w ;
 	bmpBGTable_[ arg_bmpNo ]._h = arg_h ;
@@ -190,8 +190,8 @@ int BackGround::setBmpData(
 		int arg_bmpNo ,
 		int arg_anchor ,
 		float arg_x , float arg_y ,
-		int arg_u , int arg_v ,
-		int arg_w , int arg_h ,
+		float arg_u , float arg_v ,
+		float arg_w , float arg_h ,
 		float arg_scaleX , float arg_scaleY ,
 		int arg_alpha ,
 		float arg_degree
